+++ lib/augmented/exceptions.rb
@@ -1,9 +1,11 @@
 require 'augmented/exceptions/chain'
 require 'augmented/exceptions/detailed'
+require 'augmented/exceptions/hashable'
 
 module Augmented
   module Exceptions
     include Chain
     include Detailed
+    include Hashable
   end
 end