+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## [Unreleased]
 
-- Added `Exception#chain`
 - Added `Exception#details`, `Exception#details=`, `Exception#detailed`
+- Added `Exception#chain`
+- Added `Exception#to_h`
 
 ## [0.2.3] - 2021-05-29